delete pot, then delete photo

+++ app/Http/Controllers/Admin/PotController.php
@@ -35,7 +35,7 @@ class PotController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -64,7 +64,7 @@ class PotController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -75,7 +75,7 @@ class PotController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
@@ -88,8 +88,8 @@ class PotController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -110,14 +110,21 @@ class PotController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $pot = Pot::findOrFail($id);
-        $pot->delete();
+        $photo = Pot::where('id', $id)->first();
+        $file = "storage/$photo->pot_image";
+        if (file_exists($file)) {
+            unlink($file);
+            $pot = Pot::findOrFail($id);
+            $pot->delete();
 
-        return redirect()->back()->with('success', 'Pot is successfully deleted');
+            return redirect()->back()->with('success', 'Pot is successfully deleted');
+        } else {
+            return redirect()->back()->with('success', 'Gagal menghapus, file foto tidak ditemukan');
+        }
     }
 }
